:recycle: Update paragraphs for tests and move stub to setup
